test(server): split oversized history tests
